+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+
+## 0.1.1
+
+- Fix security issue: JWS with algorithm `none` was previously verified, 
+now you can specify which algorithms are allowed and by default `none` is 
+not allowed.  
+
 ## 0.1.0
 
 - Initial version